# docs/api/v1-openapi.yaml
openapi: 3.0.3
info:
  title: Naledi Network API (v1)
  version: "1.2.0"
  license:
    name: Proprietary (Bana Pele A13 demo)
  description: |
    REST API for partner backends to reach the Naledi Beckn network. Each call
    is translated into signed Beckn messages by the official onix adapters.

    **Flow** (each step uses the `transactionId` returned in step 1):

    1. `POST /v1/search`: find providers for a need
    2. `GET /v1/results/{transactionId}`: see providers and offers
    3. `POST /v1/provider/offer`: provider sends an offer (optional)
    4. `POST /v1/select`: practitioner accepts a provider
    5. `POST /v1/provider/decision`: provider accepts or declines
    6. `POST /v1/confirm`: practitioner confirms support received
    7. `POST /v1/provider/complete`: provider marks it delivered

    Use `GET /v1/status/{transactionId}` at any time to track progress.

    **Coordination (UC1):** `GET /v1/commitments` is the shared view of who
    holds which need, and `/v1/subscriptions` lets an NGO portal be notified
    of new matches, requests and status changes.

    **Authentication:** header `X-Api-Key` on every call except `GET /v1/health`.
    The base URL and key are shared privately.
servers:
  - url: "{baseUrl}"
    description: naledi-edge. The address is sent to you privately.
    variables:
      baseUrl:
        default: https://your-sandbox-address
        description: The base URL we send you (tunnel or Azure sandbox)
security:
  - ApiKey: []
...
paths:
...
  /v1/search:
    post:
      tags: ["1. Search"]
      operationId: startSearch
      summary: Start a search
      description: |
        Call it when the practitioner signs up or asks for help. Sends a Beckn
        `discover` into the network with only the fields she agreed to share.
        Keep the returned `transactionId` for all later calls.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchRequest' }
            example: { practitionerId: prac_naledi_mahlangu, needType: registration, region: Alexandra, tier: Pre-Bronze, children: 28 }
      responses:
        '202':
          description: Search sent. Results arrive shortly (see Get results).
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId: { type: string, format: uuid }
                  status: { type: string, enum: [searching] }
                  epoch: { type: integer }
              example: { transactionId: 34d36803-11dd-44b8-9091-c466f85d20f0, status: searching, epoch: 3 }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/NetworkRejected' }
...
  /v1/select:
    post:
      tags: ["4. Select"]
      operationId: selectProvider
      summary: Accept a provider
      description: |
        Sends Beckn `select` then `init`. The status becomes `pending` until
        the provider decides. Repeating it for the same provider is
        safe (200, `idempotent: true`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId, practitionerId, needType, providerId]
              properties:
                transactionId: { type: string }
                practitionerId: { type: string }
                needType: { $ref: '#/components/schemas/NeedType' }
                providerId: { type: string }
            example: { transactionId: 34d36803-11dd-44b8-9091-c466f85d20f0, practitionerId: prac_naledi_mahlangu, needType: registration, providerId: provider-wehelp }
      responses:
        '202':
          description: Request sent to the provider
          content:
            application/json:
              schema: { type: object, properties: { transactionId: { type: string }, status: { type: string, enum: [pending] } } }
              example: { transactionId: 34d36803-11dd-44b8-9091-c466f85d20f0, status: pending }
        '200': { $ref: '#/components/responses/Idempotent' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UnknownTransaction' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/NetworkRejected' }
...
  /v1/confirm:
    post:
      tags: ["6. Confirm"]
      operationId: confirmSupport
      summary: Confirm support received
      description: |
        Needs a `reserved` request. Sends Beckn `confirm`.
        The status becomes `confirming` until the provider completes.
        Repeating it is safe. The `note` stays in our record and does not
        travel through the network.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId]
              properties:
                transactionId: { type: string }
                practitionerId: { type: string }
                needType: { $ref: '#/components/schemas/NeedType' }
                providerId: { type: string }
                note: { type: string }
            example: { transactionId: 34d36803-11dd-44b8-9091-c466f85d20f0, note: Registration documents submitted }
      responses:
        '202':
          description: Confirm sent to the provider
          content:
            application/json:
              schema: { type: object, properties: { transactionId: { type: string }, status: { type: string, enum: [confirming] } } }
              example: { transactionId: 34d36803-11dd-44b8-9091-c466f85d20f0, status: confirming }
        '200': { $ref: '#/components/responses/Idempotent' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UnknownTransaction' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/NetworkRejected' }
  /v1/withdraw:
    post:
      tags: ["Withdraw"]
      operationId: withdrawRequest
      summary: Withdraw a request
      description: |
        The practitioner no longer needs the help. If a provider holds or is
        deciding on the request, sends Beckn `cancel`; the status becomes
        `withdrawing`, then `withdrawn`, and the provider and coach are released.
        Before any provider is involved it is closed at once (200).
        A fulfilled request cannot be withdrawn (409 `already_fulfilled`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId]
              properties:
                transactionId: { type: string }
                reason: { type: string }
            example: { transactionId: 34d36803-11dd-44b8-9091-c466f85d20f0, reason: No longer needed }
      responses:
        '202':
          description: Cancel sent to the provider
          content:
            application/json:
              schema: { type: object, properties: { transactionId: { type: string }, status: { type: string, enum: [withdrawing] } } }
        '200':
          description: Closed (nobody held it yet), or already withdrawn
          content:
            application/json:
              schema: { type: object, properties: { transactionId: { type: string }, status: { type: string, enum: [withdrawn, withdrawing] }, idempotent: { type: boolean } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UnknownTransaction' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/NetworkRejected' }
...
  /v1/commitments:
    get:
      tags: ["Coordination"]
      operationId: listCommitments
      summary: List commitments (who holds which need)
      description: |
        The per-Naledi shared view: every need, its status and the provider
        holding it, so NGOs can see what is already being handled and avoid
        duplicate work. All filters are optional and combine.
      parameters:
        - { name: practitionerId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [open, reserved, fulfilled] } }
        - { name: providerId, in: query, schema: { type: string } }
        - { name: needType, in: query, schema: { $ref: '#/components/schemas/NeedType' } }
        - { name: coachId, in: query, schema: { type: string }, description: Only needs assigned to this coach (coach app). }
      responses:
        '200':
          description: Matching commitments
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  commitments: { type: array, items: { $ref: '#/components/schemas/Commitment' } }
              example:
                count: 1
                commitments:
                  - { needId: "prac_naledi_mahlangu:registration", practitionerId: prac_naledi_mahlangu, needType: registration, region: Alexandra, status: reserved, providerId: provider-wehelp, providerName: WeHelp, coachId: coach_thabo_nkosi, transactionId: 34d36803-11dd-44b8-9091-c466f85d20f0, updatedAt: "2026-10-06T08:12:40.000Z", awaitingDecision: null }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/NetworkOrProviderSideError' }
  /v1/subscriptions:
    get:
      tags: ["Coordination"]
      operationId: listSubscriptions
      summary: List NGO subscriptions
      description: Secrets are never returned here.
      responses:
        '200':
          description: All subscriptions
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Subscription' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: ["Coordination"]
      operationId: createSubscription
      summary: Subscribe an NGO to network events
      description: |
        We POST each matching event to `url`, with the header
        `X-Api-Key: <secret>`, in the same format as the partner events
        (see `docs/PARTNER-CONTRACT.md`). Failed deliveries are retried with
        backoff, in order, per subscription; answer with any 2xx. Ignore an
        `eventId` you have already handled. The `secret` is returned only in
        this response: send your own, or keep the generated one.
        Empty filter lists mean "all".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
                name: { type: string }
                secret: { type: string, description: Sent back in X-Api-Key on every delivery. Generated if missing. }
                providerIds: { type: array, items: { type: string } }
                needTypes: { type: array, items: { $ref: '#/components/schemas/NeedType' } }
                regions: { type: array, items: { type: string } }
                events:
                  type: array
                  items: { type: string, enum: [practitioner.matched, request.received, status.changed] }
                  description: Defaults to practitioner.matched and request.received.
            example: { name: WeHelp portal, url: "https://ngo-portal.example/network/events", providerIds: [provider-wehelp], needTypes: [registration], events: [practitioner.matched, request.received, status.changed] }
      responses:
        '201':
          description: Created. The only response that includes the secret.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Subscription' }
                  - { type: object, properties: { secret: { type: string } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1/admin/reset:
    post:
      tags: ["Admin"]
      operationId: resetSandbox
      summary: Reset sandbox state
      description: Increases `epoch`. When an event carries a new epoch, drop the network state you hold. The log is kept.
      responses:
        '200':
          description: Reset
          content:
            application/json:
              schema: { type: object, properties: { epoch: { type: integer }, providerSide: { type: string } } }
              example: { epoch: 4, providerSide: ok }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Sent to you privately. We send the same header when we call your /network/* endpoints.
...
  responses:
...
    InvalidRequest:
      description: invalid_request (a required field is missing, or the body is not JSON)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...